[nbs_load_actor] nbs load test lib was moved into ydb/core/cloud
