+++ ydb/core/nbs/cloud/blockstore/tools/testing/loadtest/lib/public.h
@@ -2,7 +2,6 @@
 
 #include <ydb/core/nbs/cloud/blockstore/libs/common/block_range.h>
 #include <ydb/core/nbs/cloud/storage/core/libs/common/error.h>
-#include <library/cpp/threading/future/core/future.h>
 
 #include <memory>
