+++ ydb/core/nbs/cloud/blockstore/tools/testing/loadtest/lib/range_allocator.h
@@ -3,8 +3,7 @@
 #include "public.h"
 
 #include <ydb/core/nbs/cloud/blockstore/libs/common/block_range.h>
-
-#include <ydb/core/protos/nbs2_load.pb.h>
+#include <ydb/core/nbs/cloud/blockstore/tools/testing/loadtest/lib/protos/nbs2_load.pb.h>
 
 #include <util/generic/vector.h>
