+++ ydb/core/nbs/cloud/blockstore/tools/testing/loadtest/lib/request_generator.h
@@ -1,13 +1,11 @@
 #pragma once
 
 #include "public.h"
+#include <ydb/core/nbs/cloud/blockstore/tools/testing/loadtest/lib/protos/nbs2_load.pb.h>
 
 #include <ydb/core/nbs/cloud/blockstore/libs/common/block_range.h>
 #include <ydb/core/nbs/cloud/blockstore/libs/diagnostics/public.h>
 #include <ydb/core/nbs/cloud/blockstore/libs/service/request.h>
-
-#include <ydb/core/protos/load_test.pb.h>
-#include <ydb/core/protos/nbs2_load.pb.h>
 
 #include <util/datetime/base.h>
 #include <util/generic/string.h>
