+++ ydb/core/nbs/cloud/blockstore/tools/testing/loadtest/ya.make
@@ -1,0 +1,3 @@
+RECURSE(
+    lib
+)
