+++ ydb/core/nbs/cloud/blockstore/tools/testing/ya.make
@@ -1,0 +1,3 @@
+RECURSE(
+    loadtest
+)
