+++ ydb/core/nbs/cloud/blockstore/tools/ya.make
@@ -1,0 +1,3 @@
+RECURSE(
+    testing
+)
