+++ ydb/core/nbs/cloud/blockstore/ya.make
@@ -1,4 +1,5 @@
 RECURSE(
     libs
     public
+    tools
 )
